Good stuff working on raspberry pi!
A real and somewhat polished version of this that runs on my raspberry pi
and outputs video to FadeCandy at a pretty solid framerate!
Everything's still hard-coded to 40x10 pixels. Needs lots of work to make it
generic, and ofxOPC needs a lot of tweaking to be more faithful to OPC.